# open-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/lessons": {
      "get": {
        "operationId": "LessonsController_lessonList",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Lesson list here",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedException"
                }
              }
            }
          }
        },
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "Token auth": []
          }
        ]
      },
      "post": {
        "operationId": "LessonsController_addLesson",
        "summary": "add lesson:",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddLessonAnswerDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedException"
                }
              }
            }
          }
        },
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "Token auth": []
          }
        ]
      }
    },
    "/api/lessons/{id}/evaluations": {
      "post": {
        "operationId": "LessonsController_addEvaluation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEvaluationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddEvaluationAnswerDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedException"
                }
              }
            }
          }
        },
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "Token auth": []
          }
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_pupilList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddPupilAnswerDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Token auth": []
          }
        ]
      },
      "post": {
        "operationId": "UsersController_addPupil",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddPupilAnswerDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Token auth": []
          }
        ]
      }
    },
    "/api/auth": {
      "post": {
        "operationId": "AuthController_reg",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "token created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthClientAnswerDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "NewDisk Pupil API's",
    "description": "List of API's available",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "Token auth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    },
    "schemas": {
      "UnauthorizedException": {
        "type": "object",
        "properties": {}
      },
      "LessonEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Музыка"
          },
          "code": {
            "type": "string",
            "example": "music"
          }
        },
        "required": [
          "id",
          "name",
          "code"
        ]
      },
      "AddLessonDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "music"
          },
          "name": {
            "type": "string",
            "example": "Музыка"
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "AddLessonAnswerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "code": {
            "type": "string",
            "example": "music"
          },
          "name": {
            "type": "string",
            "example": "Музыка"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ]
      },
      "AddEvaluationDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "1"
          },
          "score": {
            "type": "string",
            "example": "56"
          }
        },
        "required": [
          "user_id",
          "score"
        ]
      },
      "AddEvaluationAnswerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3"
          },
          "user_id": {
            "type": "string",
            "example": "1"
          },
          "score": {
            "type": "string",
            "example": "56"
          }
        },
        "required": [
          "id",
          "user_id",
          "score"
        ]
      },
      "AddPupilAnswerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Джонни"
          },
          "email": {
            "type": "string",
            "example": "silverhand@mail.com"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Джонни"
          },
          "email": {
            "type": "string",
            "example": "silverhand@mail.com"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "AuthClientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "name1"
          },
          "pass": {
            "type": "string",
            "example": "pass1"
          }
        },
        "required": [
          "name",
          "pass"
        ]
      },
      "AuthClientAnswerDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "57e0209fedeb4cfabaa26d1acf57af51"
          }
        },
        "required": [
          "token"
        ]
      }
    }
  }
}